Remove deprecated API calls where possible.
