+++ idea-p4server/api/src/main/java/net/groboclown/p4/server/api/ApplicationPasswordRegistry.java
@@ -27,6 +27,24 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.jetbrains.concurrency.Promise;
 
+import java.util.Arrays;
+
+
+/*
+ * FIXME
+ * Use application services or extensions instead of ApplicationComponent, because
+ * if you register a class as an application component it will be loaded, its instance will be created and
+ * {@link #initComponent()} methods will be called each time IDE is started even if user doesn't use any feature of your
+ * plugin. So consider using specific extensions instead to ensure that the plugin will not impact IDE performance until user calls its
+ * actions explicitly.
+ *
+ * Instead of {@link #initComponent()} please use {@link com.intellij.util.messages.MessageBus} and corresponding topics.
+ * Instead of {@link #disposeComponent()} please use {@link com.intellij.openapi.Disposable}.
+ *
+ * If for some reasons replacing {@link #disposeComponent()} / {@link #initComponent()} is not a option, {@link BaseComponent} can be extended.
+ */
+
+
 /**
  * Maintains the registry of plaintext passwords (either stored in files or entered by the user).
  * Passwords are stored in such a way that there may be a delay in fetching the actual value,
@@ -39,19 +57,17 @@ public abstract class ApplicationPasswordRegistry
         implements ApplicationComponent {
     private static final Logger LOG = Logger.getInstance(ApplicationPasswordRegistry.class);
 
+    public static final Class<ApplicationPasswordRegistry> COMPONENT_CLASS = ApplicationPasswordRegistry.class;
     public static final String COMPONENT_NAME = "p4ic4idea:pr";
 
     @NotNull
     public static ApplicationPasswordRegistry getInstance() {
-        return (ApplicationPasswordRegistry) ApplicationManager.getApplication().getComponent(COMPONENT_NAME);
+        return ApplicationManager.getApplication().getComponent(COMPONENT_CLASS);
     }
 
     /**
      * Stores the given password in the registry, associated with the config.  After being
      * called, the password will be blanked out.
-     *
-     * @param config
-     * @param password
      */
     public final void store(@NotNull ServerConfig config, @NotNull char[] password, boolean inMemoryOnly) {
         if (LOG.isDebugEnabled()) {
@@ -59,9 +75,7 @@ public abstract class ApplicationPasswordRegistry
         }
         final CredentialAttributes attr = getCredentialAttributes(config, inMemoryOnly);
         PasswordSafe.getInstance().set(attr, new Credentials(config.getUsername(), password));
-        for (int i = 0; i < password.length; i++) {
-            password[i] = 0;
-        }
+        Arrays.fill(password, (char) 0);
     }
 
     /**
@@ -110,6 +124,7 @@ public abstract class ApplicationPasswordRegistry
         final CredentialAttributes attr = getCredentialAttributes(config, true);
         return PasswordSafe.getInstance().getAsync(attr)
                 .then((c) -> c == null ? null : c.getPassword())
+                // should use onProcessed, but that's a higher API version.
                 .processed((p) -> {
                     if (LOG.isDebugEnabled()) {
                         LOG.debug("Fetched password for config " + config.getServerName());
