+++ idea-p4server/impl/src/test/java/net/groboclown/p4/server/impl/cache/store/ProjectCacheStoreTest.java
@@ -17,12 +17,11 @@ import com.intellij.util.xmlb.XmlSerializer;
 import net.groboclown.idea.extensions.IdeaLightweightExtension;
 import net.groboclown.idea.extensions.TemporaryFolder;
 import net.groboclown.idea.extensions.TemporaryFolderExtension;
-import net.groboclown.idea.mock.MockVirtualFileSystem;
 import net.groboclown.p4.server.api.ClientServerRef;
 import net.groboclown.p4.server.api.P4ServerName;
 import net.groboclown.p4.server.api.commands.changelist.MoveFilesToChangelistAction;
 import net.groboclown.p4.server.impl.values.P4ChangelistIdImpl;
-import net.groboclown.p4.server.todo.util.mock.MockFilePath;
+import net.groboclown.idea.altmock.MockFilePath;
 import org.jdom.Element;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
