+++ idea-test-core/src/main/java/net/groboclown/idea/altmock/MockFileType.java
@@ -1,0 +1,66 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package net.groboclown.idea.altmock;
+
+import com.intellij.openapi.fileTypes.FileType;
+import com.intellij.openapi.vfs.VirtualFile;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+import javax.swing.*;
+
+public class MockFileType implements FileType {
+    public static final FileType MOCK = new MockFileType();
+
+    @NotNull
+    @Override
+    public String getName() {
+        return "mock";
+    }
+
+    @NotNull
+    @Override
+    public String getDescription() {
+        return "mock";
+    }
+
+    @NotNull
+    @Override
+    public String getDefaultExtension() {
+        return ".mock";
+    }
+
+    @Nullable
+    @Override
+    public Icon getIcon() {
+        return null;
+    }
+
+    @Override
+    public boolean isBinary() {
+        return false;
+    }
+
+    @Override
+    public boolean isReadOnly() {
+        return false;
+    }
+
+    @Nullable
+    @Override
+    public String getCharset(@NotNull VirtualFile file, @NotNull byte[] content) {
+        return null;
+    }
+}
