+++ plugin-v3/src/main/resources/META-INF/plugin.xml
@@ -7,7 +7,7 @@
     <category>VCS Integration</category>
     <change-notes><![CDATA[
     <ul>
-      <li><em>0.10.8</em><ul>
+      <li><em>0.10.9</em><ul>
         <li>Fixed API usage to support future IDE versions.</li>
       </ul></li>
       <li><em>0.10.8</em><ul>
